feat: Improve menu, routing, Sidebar layout
Make a condition that the second-level menus are hidden and only the active one is open. Make a selection of the selected menu item of the third level. Make the function of opening the menu of the second level. Create a Search Page. Set up 2-level routing. Sidebar layout. Create a Page Component